fix(security): break PrismaService ↔ TenantInterceptor circular import
The previous commit added a `PrismaService` constructor injection to
`TenantInterceptor`, but `PrismaService.runWithRlsTenant` already
imports `getCurrentTenantId` from `tenant.interceptor.ts` — so loading
`PrismaService` first now hits the interceptor module before its
own decorators have run. CI smoke job caught it:

    ReferenceError: Cannot access 'PrismaService' before initialization
        at src/core/multi-tenancy/tenant.interceptor.ts:57:8

Fix: move the `AsyncLocalStorage` container + `getCurrentTenantId` /
`runWithTenant` to a leaf module `tenant-context.ts` that imports
nothing from the rest of the project. Both `PrismaService` and
`TenantInterceptor` consume it without forming a cycle. The
interceptor re-exports the two functions so existing callers (Example
controller, scaffold-module-planner, the `tenant-interceptor` story
test, et al.) keep their import paths.

Also adapts `tests/better-auth-session-middleware.e2e-spec.ts` to the
new "header pointing at a tenant the user has no ACTIVE membership in
→ 403" rule: the test now provisions the test tenant + sets it as the
user's primary so the resolver short-circuits the membership lookup
(this is the same pattern `files-persistence.e2e-spec.ts` and
`me-tenants-self-service.e2e-spec.ts` already use).
